prompt dialogue: re-impl as standalone dialogue
Previously, the implementation of `PromptDialogueState` had a
`ButtonDialogueState`. This implementation was poor because the button
dialogue was unaware of the input element. `PromptDialogueState` has
been reimplemented as a fully standalone dialogue that manages its own
state.

Now the dialogue cursor correctly moves between the input and the
buttons. However, the input does not change visually when it is
(un)focused.